// dashboard/types/generated-types.ts
export interface definitions {
  LastBlock: {
    /**
     * Format: integer
     * @description Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    /** Format: integer */
    blockNumber?: number;
  };
  Event: {
    /**
     * Format: integer
     * @description Note:
     * This is a Primary Key.<pk/>
     */
    id: number;
    /** Format: bigint */
    blockNumber?: number;
    /** Format: character varying */
    txHash?: string;
    /** Format: integer */
    txIndex?: number;
    /** Format: integer */
    logIndex?: number;
    /** Format: boolean */
    removed?: boolean;
    /** Format: bigint */
    timestamp?: number;
    /** Format: text */
    type?: string;
    /** Format: jsonb */
    data?: unknown;
  };
}